requirements.txt
asgiref==3.5.0
crispy-bootstrap5==0.6
Django==4.0.3
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
Markdown==3.3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
